requirements.txt
Django==1.4.1
South==0.7.6
bleach==1.1.5
cssselect==0.7.1
distribute==0.6.24
html5lib==0.95
lxml==2.3.6
psycopg2==2.4.5
pyquery==1.2.1
python-dateutil==2.1
requests==0.14.1
simplejson==2.6.2
six==1.2.0
wsgiref==0.1.2
